클래스 나누기
