+++ src/test/java/kr/codesquad/view/InputTest.java
@@ -1,5 +1,6 @@
 package kr.codesquad.view;
 
+import kr.codesquad.domain.Validator;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
@@ -9,10 +10,10 @@ public class InputTest {
     @Test
     @DisplayName("글자 크기 실패 테스트")
     void NameFailTest() {
-        Input input = new Input();
+        Validator validator = new Validator();
         String[] names = {"wisdom", "a", "aa"};
 
-        boolean result = input.validateNames(names);
+        boolean result = validator.validateNames(names);
 
         assertThat(result).isFalse();
     }
@@ -20,10 +21,10 @@ public class InputTest {
     @Test
     @DisplayName("글자 크기 성공 테스트")
     void NameSuccessTest() {
-        Input input = new Input();
+        Validator validator = new Validator();
         String[] names = {"pobi", "honux", "crong", "jk"};
 
-        boolean result = input.validateNames(names);
+        boolean result = validator.validateNames(names);
 
         assertThat(result).isTrue();
     }
